sample_code/billing_and_webhooks/team_billing/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
mysql-connector-python==8.1.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-dotenv==1.0.0
sqlmodel==0.0.16
sqlalchemy==2.0.29
